Move random events, roundstart away mission loading and night shifts to per server
Meaning event hall can (and does) opt out of things.

Problem is these flags are a "presence" flag, the token appearing in a config enables it so once a config does that you cant override it per server, only ensure it's not in the base and enable it per server to allow event hall to opt out.